@@ -5,7 +5,7 @@ author: Martin Rybak <martin.rybak@gmail.com>
 homepage: https://github.com/vgventures/futuristic
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,7 +14,11 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.9.0                  # Static analysis rules
+  pedantic: ^1.11.0                  # Static analysis rules
+  
+analyzer:
+  enable-experiment:
+  - non-nullabl
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
